Modify Users and Trainers.
The registration of a user has been improved. At the time of
registration, a notification is created for the admin, which confirms
the user and an email with the confirmation is sent. The index of the
trainershas been modified so that the profile picture, the name and the
description are displayed.